Fix delete message tests
